modified user model and firebasecontroller to remove facility setter

+++ Soroban/app/src/main/java/com/example/soroban/controller/UserController.java
@@ -34,7 +34,7 @@ public class UserController {
         user.setFirstName(firstName.toString());
         user.setLastName(lastName.toString());
         user.setEmail(email.toString());
-        int phoneNumber = Integer.parseInt(number.toString().replaceAll("\\D", ""));
+        long phoneNumber = Long.parseLong(number.toString().replaceAll("\\D", ""));
         user.setPhoneNumber(phoneNumber);
     }
